requirements.txt
amqp==5.2.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
cachetools==4.2.4
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
construct==2.10.68
construct-typing==0.5.6
cryptography==41.0.7
daphne==4.0.0
Deprecated==1.2.14
dj-database-url==2.1.0
Django==5.0
django-countries==7.5.1
django-oauth-toolkit==2.3.0
django-otp==1.3.0
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Faker==21.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jsonalias==0.1.1
jwcrypto==1.5.0
kombu==5.3.4
mnemonic==0.20
msgpack==1.0.7
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
phonenumbers==8.13.26
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rfc3986==1.5.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
types-cachetools==4.2.10
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
websockets==11.0.3
wrapt==1.16.0
zope.interface==6.1
